requirements.txt
Flask
google-generativeai
tensorflow
matplotlib
Markdown
nltk
numpy
pillow
pymongo
dnspython
python-dotenv
scikit-learn
spellchecker
